requirements.txt
PyAudio
google-genai
asyncio
flet